FlaskApp/requirements.txt
bcrypt==3.1.4
cffi==1.11.5
click==6.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-MongoAlchemy==0.7.2
Flask-PyMongo==0.5.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
MongoAlchemy==0.19
Pillow==5.1.0
pycparser==2.18
pymongo==2.8.1
six==1.11.0
Werkzeug==0.14.1
WTForms==2.1
